+++ spec/data/test_config_invalid.rb
@@ -1,4 +1,4 @@
-repository_manager :type => 'Gitlab', :base_url => 'http://localhost:8080', :token =>'test_token'
+repository_manager :type => 'GitLab', :base_url => 'http://localhost:8080', :token =>'test_token'
 chef_server_url 'http://localhost:4000'
 chef_client_pem 'testclient.pem'
 chef_username 'test_user'