+++ spec/gitlab_manager_spec.rb
@@ -1,14 +1,14 @@
 require_relative 'support/spec_helper'
 
-describe GitlabManager do
+describe GitLabManager do
   describe "API" do
     let(:manager) do
       config = {:base_url => "Test url",
           :token => "token",
-          :type => "Gitlab"
+          :type => "GitLab"
         }
 
-      manager = GitlabManager.new config
+      manager = GitLabManager.new config
       manager
     end
 
@@ -25,33 +25,33 @@ describe GitlabManager do
       it "creates a valid Manager" do
         config = {:base_url => "Test url",
                   :token => "token",
-                  :type => "Gitlab"
+                  :type => "GitLab"
                 }
 
-        manager = GitlabManager.new config
+        manager = GitLabManager.new config
         manager
       end
     end
 
     context "invalid configuration" do
-      it "raises an error when Gitlab Token is not configured" do
+      it "raises an error when GitLab Token is not configured" do
         config = {:base_url => "Test url",
-                  :type => "Gitlab"
+                  :type => "GitLab"
                 }
 
         expect do
-          GitlabManager.new config
-        end.to raise_error(GitlabAccessNotConfiguredError)
+          GitLabManager.new config
+        end.to raise_error(GitLabAccessNotConfiguredError)
       end
 
-      it "raises an error when Gitlab Base Url is not configured" do
+      it "raises an error when GitLab Base Url is not configured" do
         config = {:token => "token",
-                  :type => "Gitlab"
+                  :type => "GitLab"
                 }
 
         expect do
-          GitlabManager.new config
-        end.to raise_error(GitlabAccessNotConfiguredError)
+          GitLabManager.new config
+        end.to raise_error(GitLabAccessNotConfiguredError)
       end
     end
   end