+++ CHANGELOG.md
@@ -8,7 +8,7 @@ FEATURES:
 * Add option to recursively analyze cookbook's dependencies (defaults to true)
 * Use a file to store configuration info (defaults to *${HOME}/.chef/kitchen_inspector.rb*)
 * Add option to show remarks (*--remarks*), that now are hidden by default
-* Add support for different Repository Managers (Gitlab and GitHub)
+* Add support for different Repository Managers (GitLab and GitHub)
 * Verifies that a tag on the Repository Manager matches the corresponding metadata's version
 * Add exit codes to integrate Kitchen Inspector with a CI process
 * Handle not unique projects on Repository Manager